클라이언트 인증

+++ database/seeds/DatabaseSeeder.php
@@ -70,52 +70,55 @@ class DatabaseSeeder extends Seeder
         $this->command->info('Seeded: article_tag table');
 
         /* 첨부 파일 */
-        App\Attachment::truncate();
+        if (config('database.default') != 'testing') {
+            // testing 환경이 아닐 때만...
+            App\Attachment::truncate();
 
-        if (! File::isDirectory(attachments_path())) {
-            File::makeDirectory(attachments_path(), 775, true);
-        }
+            if (! File::isDirectory(attachments_path())) {
+                File::makeDirectory(attachments_path(), 775, true);
+            }
 
-        File::cleanDirectory(attachments_path());
+            File::cleanDirectory(attachments_path());
 
-        // public/files/.gitignore 파일이 있어야 커밋할 때 빈 디렉터리를 유지할 수 있다.
-        File::put(attachments_path('.gitignore'), "*\n!.gitignore");
+            // public/files/.gitignore 파일이 있어야 커밋할 때 빈 디렉터리를 유지할 수 있다.
+            File::put(attachments_path('.gitignore'), "*\n!.gitignore");
 
-        $this->command->error(
-            'Downloading ' . $articles->count() . ' images from lorempixel. It takes time...'
-        );
-
-        $articles->each(function ($article) use ($faker) {
-            $path = $faker->image(attachments_path());
-            $filename = File::basename($path);
-            $bytes = File::size($path);
-            $mime = File::mimeType($path);
-
-            $this->command->warn("File saved: {$filename}");
-
-            $article->attachments()->save(
-                factory(App\Attachment::class)->make(compact('filename', 'bytes', 'mime'))
+            $this->command->error(
+                'Downloading ' . $articles->count() . ' images from lorempixel. It takes time...'
             );
-        });
 
-        foreach(range(1, 10) as $index) {
-            // 테스트를 위해 고아가 된 첨부파일을 만든다.
-            // 고아가 된 첨부파일 이란 article_id가 없고 생성된 지 일주일 넘은 테이블 레코드/파일를 의미한다.
-            $path = $faker->image(attachments_path());
-            $filename = File::basename($path);
-            $bytes = File::size($path);
-            $mime = File::mimeType($path);
-            $this->command->warn("File saved: {$filename}");
+            $articles->each(function ($article) use ($faker) {
+                $path = $faker->image(attachments_path());
+                $filename = File::basename($path);
+                $bytes = File::size($path);
+                $mime = File::mimeType($path);
 
-            factory(App\Attachment::class)->create([
-                'filename' => $filename,
-                'bytes' => $bytes,
-                'mime' => $mime,
-                'created_at' => $faker->dateTimeBetween('-1 months'),
-            ]);
+                $this->command->warn("File saved: {$filename}");
+
+                $article->attachments()->save(
+                    factory(App\Attachment::class)->make(compact('filename', 'bytes', 'mime'))
+                );
+            });
+
+            foreach(range(1, 10) as $index) {
+                // 테스트를 위해 고아가 된 첨부파일을 만든다.
+                // 고아가 된 첨부파일 이란 article_id가 없고 생성된 지 일주일 넘은 테이블 레코드/파일를 의미한다.
+                $path = $faker->image(attachments_path());
+                $filename = File::basename($path);
+                $bytes = File::size($path);
+                $mime = File::mimeType($path);
+                $this->command->warn("File saved: {$filename}");
+
+                factory(App\Attachment::class)->create([
+                    'filename' => $filename,
+                    'bytes' => $bytes,
+                    'mime' => $mime,
+                    'created_at' => $faker->dateTimeBetween('-1 months'),
+                ]);
+            }
+
+            $this->command->info('Seeded: attachments table and files');
         }
-
-        $this->command->info('Seeded: attachments table and files');
 
         /* 댓글 */
         $articles->each(function ($article) {
@@ -143,13 +146,17 @@ class DatabaseSeeder extends Seeder
         $this->command->info('Seeded: comments table');
 
         /* up & down 투표 */
-        $comments = App\Comment::all();
+        if (config('database.default') != 'testing') {
+            // testing 환경이 아닐 때만...
+            // Sqlite가 up, down 열에 대해 'Integrity constraint violation' 일으킴
+            $comments = App\Comment::all();
 
-        $comments->each(function ($comment) {
-            $comment->votes()->save(factory(App\Vote::class)->make());
-            $comment->votes()->save(factory(App\Vote::class)->make());
-            $comment->votes()->save(factory(App\Vote::class)->make());
-        });
+            $comments->each(function ($comment) {
+                $comment->votes()->save(factory(App\Vote::class)->make());
+                $comment->votes()->save(factory(App\Vote::class)->make());
+                $comment->votes()->save(factory(App\Vote::class)->make());
+            });
+        }
 
         $this->command->info('Seeded: votes table');
     }
